Change source file for ruby binary from symlink to file
Unix-style symlinks are treated as regular files on Windows with text contents
of the link target. This is not buildable rust code.

This commit refactors the ruby driver to be library code in artichoke-frontend
that is delegated to by a small shim in main in both artichoke.rs and ruby.rs.
