+++ artichoke-frontend/src/lib.rs
@@ -7,3 +7,4 @@
 
 pub mod parser;
 pub mod repl;
+pub mod ruby;
